Wrap and annotate in FuseOpsByPattern optionally
